Moving things around to accommodate cli version
Still need to test pyinstaller on mac/win
