--- pdfstitcher/__main__.py
@@ -1,3 +1,0 @@
-from pdfstitcher.app import main
-
-main()
