+++ pdfstitcher/gui/dialogs.py
