+++ pdfstitcher/gui/layers_tab.py
