+++ pdfstitcher/processing/pagetiler.py
@@ -147,7 +147,6 @@ class PageTiler(ProcessingBase):
             bbox = copy.copy(in_doc_page.trimbox)
 
         # set the trim box to cut off content if requested
-        # TODO: Consider moving this to layout functions
         if self.p["actually_trim"]:
             page_trim = self._get_trim(page_uu)
             # things get tricky if there's rotation, because the user sees top/bottom as right/left
@@ -658,145 +657,3 @@ def main(args: argparse.Namespace) -> None:
         print(_("Successfully written to") + " " + args.output)
     else:
         print(_("Something went wrong"))
-
-
-def parse_arguments() -> argparse.Namespace:
-    """
-    Helper function to parse command line arguments.
-    """
-    parser = argparse.ArgumentParser(
-        description=_("Tile PDF pages into one document."),
-        epilog=_("Note: If both rows and columns are specified, rows are ignored.")
-        + " "
-        + _("To insert a blank page, include a zero in the page list."),
-    )
-
-    # Required arguments
-    parser.add_argument(
-        "input",
-        help=_("Input filename (pdf)"),
-    )
-    parser.add_argument(
-        "output",
-        help=_("Output filename (pdf)"),
-    )
-    group = parser.add_argument_group(
-        _("Grid layout"), _("Number of Columns") + " " + _("OR Number of Rows")
-    )
-    m_group = group.add_mutually_exclusive_group(required=True)
-    m_group.add_argument(
-        "-r",
-        "--rows",
-        type=int,
-        help=_("Number of rows in tiled grid."),
-    )
-    m_group.add_argument(
-        "-c",
-        "--columns",
-        type=int,
-        help=_("Number of columns in tiled grid."),
-    )
-
-    parser.add_argument(
-        "-p",
-        "--pages",
-        help=_(
-            "Pages to tile. May be range or list (e.g. 1-5 or 1,3,5-7, etc). Default: entire document."
-        ),
-    )
-    parser.add_argument(
-        "-u",
-        "--units",
-        choices=[_("in"), _("cm")],
-        default=_("in"),
-        help=_("Units for margin and trim values."),
-    )
-    parser.add_argument(
-        "-m",
-        "--margin",
-        help=_("Margin size in selected units."),
-    )
-    parser.add_argument(
-        "-t",
-        "--trim",
-        default="0,0,0,0",
-        help=_("Amount to trim from edges in selected units")
-        + " "
-        + _("given as left,right,top,bottom (e.g. 0.5,0,0.5,0 would trim 0.5 from left and top)"),
-    )
-    parser.add_argument(
-        "-R",
-        "--rotate",
-        type=int,
-        default=0,
-        choices=[0, 90, 180, 270],
-        help=_("Rotate pages"),
-    )
-    parser.add_argument(
-        "--col-major",
-        type=bool,
-        default=False,
-        help=_("Fill columns before rows (default is rows first)"),
-    )
-    parser.add_argument(
-        "--right-to-left",
-        type=bool,
-        default=False,
-        help=_("Fill columns right to left (default is left to right)"),
-    )
-    parser.add_argument(
-        "--bottom-to-top",
-        type=bool,
-        default=False,
-        help=_("Fill rows bottom to top (default is top to bottom)"),
-    )
-    parser.add_argument(
-        "--target-height",
-        type=float,
-        help=_("Height of output document in selected units.")
-        + " "
-        + _("Caution: results in scaling of pages"),
-        default=None,
-    )
-    parser.add_argument(
-        "--target-width",
-        type=float,
-        help=_("Width of output document in selected units.")
-        + " "
-        + _("Caution: results in scaling of pages"),
-        default=None,
-    )
-    parser.add_argument(
-        "--trimbox-to-mediabox",
-        action="store_true",
-        help=_("Override trimbox with mediabox"),
-        default=False,
-    )
-    parser.add_argument(
-        "--actually-trim",
-        action="store_true",
-        help=_("Actually trim the pages (default is overlap)"),
-        default=False,
-    )
-
-    args = parser.parse_args()
-
-    # validate the trim values
-    trim = [utils.txt_to_float(t) for t in args.trim.split(",")]
-    if len(trim) == 1:
-        args.trim = [trim[0], trim[0], trim[0], trim[0]]
-    elif len(trim) == 2:
-        args.trim = [trim[0], trim[0], trim[1], trim[1]]
-    elif len(trim) == 4:
-        args.trim = trim
-    else:
-        print(_("Invalid trim value specified, ignoring"))
-        args.trim = [0, 0, 0, 0]
-
-    return args
-
-
-if __name__ == "__main__":
-    utils.setup_locale()
-    args = parse_arguments()
-    main(args)
